Include HubSpot <script> block on all pages
As noted in the comments, refer to the Minimal Mistakes theme's
documentation for custom footers and generic script values. I wish we
could specify "async defer" for after_footer_scripts, because this
change doesn't exactly match what HubSpot requests. I can either:
1. Include the <script> right before </body> without "async defer"
OR
2. Include the "async defer" but put the <script> in the footer div,
   which comes just before the </body>

My suspicion is that "async defer" is more important for UX than
including at the end of the body, but I could be wrong.